Избавляемся от предупреждений на 6.21

+++ sources/Device/src/Hardware/Modules/BH1750/BH1750.cpp
@@ -48,7 +48,7 @@ bool BH1750::GetMeasure(float *illumination)
         return false;
     }
 
-    timeNext += TIME_MEASURE + (std::rand() % 100);
+    timeNext += TIME_MEASURE + (uint)(std::rand() % 100);
 
 #ifdef IN_MODE_TEST
